Add GitHub Action
Because I did the back-and-forth of setting up the action and
getting it to work right in the main branch, I decided to go ahead
and force push to main to get rid of some of the intermediate commits.
Apologies for any inconvenience and I'll try to avoid that in future.